return ErrFailure when error callback fires
The XHR API does not actually pass an error object to the callback, but
a progress event. Furthermore, said event is devoid of any information
about the nature of the error, and there is no other way of getting to
the error, either. And because the XHR API is asynchronous, we cannot
catch the exception either.

Closes gh-1