don't sanitize strings, always deal with binary data
In Go, strings are usually considered immutable byte slices, and their
encoding doesn't matter, at least not when transmitting them as data.

JavaScript has a different idea, which leads to strings being sanitized
so that they contain valid UTF-8, by replacing invalid byte sequences
with replacement runes.

Since we're writing this package primarily for Go programmers, we want
it to behave as close to "real" Go as possible. That's why the top level
Send function is changed to receive binary data and return []byte, and
why Request.Send is changed to never pass strings to JS, only []byte.

This makes it impossible for someone to actually make use of
aforementioned sanitization for outgoing strings, but there's no reason
XHR itself should be doing it, and it can be solved in pure Go with a
loop.

See the discussion at https://github.com/dominikh/go-js-xhr/issues/3 for
more information.